requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg[binary,pool]==3.2.1
pydantic==2.9.2
jinja2==3.1.4
python-multipart==0.0.9
pyyaml==6.0.2
